pyproject.toml
[project]
name = "biopolymerskg"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "neo4j>=6.0.2",
    "pandas>=2.3.3",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "tqdm>=4.67.1",
]

[tool.pyright]
venvPath = "."
venv = ".venv"
